# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Car Dealership API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Car {
  id: number;
  brand: string;
  model: string;
  year: number;
  price: number;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  engine?: string | null;
  /** @nullable */
  transmission?: string | null;
  /** @nullable */
  fuelType?: string | null;
  /** @nullable */
  seats?: number | null;
  /** @nullable */
  color?: string | null;
  /** @nullable */
  mileage?: number | null;
  available: boolean;
  featured: boolean;
  images: string[];
  createdAt: string;
  updatedAt: string;
}

export interface CarInput {
  /** @minLength 1 */
  brand: string;
  /** @minLength 1 */
  model: string;
  year: number;
  price: number;
  description?: string;
  engine?: string;
  transmission?: string;
  fuelType?: string;
  seats?: number;
  color?: string;
  mileage?: number;
  available?: boolean;
  featured?: boolean;
  images?: string[];
}

export interface CarUpdate {
  brand?: string;
  model?: string;
  year?: number;
  price?: number;
  description?: string;
  engine?: string;
  transmission?: string;
  fuelType?: string;
  seats?: number;
  color?: string;
  mileage?: number;
  available?: boolean;
  featured?: boolean;
  images?: string[];
}

export interface BrandCount {
  brand: string;
  count: number;
}

export interface CarStats {
  total: number;
  available: number;
  featured: number;
  byBrand: BrandCount[];
}

export interface LoginCredentials {
  /** @minLength 1 */
  username: string;
  /** @minLength 1 */
  password: string;
}

export interface AdminUser {
  id: number;
  username: string;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  admin: AdminUser;
}

export interface ContactMessage {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  email: string;
  phone?: string;
  subject?: string;
  /** @minLength 1 */
  message: string;
}

export interface ContactResult {
  success: boolean;
  message: string;
}

export type ListCarsParams = {
brand?: string;
featured?: boolean;
available?: boolean;
};