# lib/api-zod/src/generated/types/authResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Car Dealership API
 * OpenAPI spec version: 0.1.0
 */
import type { AdminUser } from './adminUser';

export interface AuthResponse {
  token: string;
  admin: AdminUser;
}