# lib/api-zod/src/generated/types/car.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Car Dealership API
 * OpenAPI spec version: 0.1.0
 */

export interface Car {
  id: number;
  brand: string;
  model: string;
  year: number;
  price: number;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  engine?: string | null;
  /** @nullable */
  transmission?: string | null;
  /** @nullable */
  fuelType?: string | null;
  /** @nullable */
  seats?: number | null;
  /** @nullable */
  color?: string | null;
  /** @nullable */
  mileage?: number | null;
  available: boolean;
  featured: boolean;
  images: string[];
  createdAt: string;
  updatedAt: string;
}