# lib/api-zod/src/generated/types/carStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Car Dealership API
 * OpenAPI spec version: 0.1.0
 */
import type { BrandCount } from './brandCount';

export interface CarStats {
  total: number;
  available: number;
  featured: number;
  byBrand: BrandCount[];
}