# lib/api-zod/src/generated/types/carUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Car Dealership API
 * OpenAPI spec version: 0.1.0
 */

export interface CarUpdate {
  brand?: string;
  model?: string;
  year?: number;
  price?: number;
  description?: string;
  engine?: string;
  transmission?: string;
  fuelType?: string;
  seats?: number;
  color?: string;
  mileage?: number;
  available?: boolean;
  featured?: boolean;
  images?: string[];
}