# lib/api-zod/src/generated/types/contactMessage.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Car Dealership API
 * OpenAPI spec version: 0.1.0
 */

export interface ContactMessage {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  email: string;
  phone?: string;
  subject?: string;
  /** @minLength 1 */
  message: string;
}