# lib/api-zod/src/generated/types/loginCredentials.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Car Dealership API
 * OpenAPI spec version: 0.1.0
 */

export interface LoginCredentials {
  /** @minLength 1 */
  username: string;
  /** @minLength 1 */
  password: string;
}